Fix Linux compilation errors in media_io files
- Fix implicit integer conversion in high_bitdepth_support.cpp
- Fix size_t to double conversion warnings in color_management.cpp
- Add explicit static_cast conversions for strict compiler warnings
- Mark unused parameters with [[maybe_unused]] attribute
- Fix matrix indexing with proper size_t casting

+++ src/media_io/src/high_bitdepth_support.cpp
@@ -801,7 +801,7 @@ uint16_t scaleBitDepth(uint16_t value, uint8_t from_bits, uint8_t to_bits) {
         return value >> (from_bits - to_bits);
     } else {
         // Increasing bit depth
-        return value << (to_bits - from_bits);
+        return static_cast<uint16_t>(value << (to_bits - from_bits));
     }
 }
